Known and expected package failures
There are a bunch of expected errors, that will and should fail with the
compiler. One case is that the package was removed 404, and another
where a tag was changed and the hash is wrong. Those are added to a list
of `errorPackages` and their failures will be snapshot tested.